+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.2.23] - 2022-10-22
+### Added
+- Default defense die per weapon.
+- Jab damage formula autocalculation.
+- Min strength weapon requirement (currently for display only).
+- Registered Gentium font with prosemirror editor.
+### Changed
+- Updated weapons compendium.
+- Redesigned various item sheets.
+- Item text editor uses prosemirror.
+
 ## [0.2.22] - 2022-10-19
 ### Added
 - Compendium items.